- Fixed some parameters in CFlux constructor method - Changed PhenArray parameters from arrays to lists - Changed Output_Map_Init to initialize lists instead of arrays

+++ BGC_Java/src/classes/CFlux.java
@@ -134,8 +134,8 @@ public class CFlux {
 	public double leafc_storage_to_leafc_transfer; /* (kgC/m2/d) */
 	public double frootc_storage_to_frootc_transfer; /* (kgC/m2/d) */
 	public double livestemc_storage_to_livestemc_transfer; /* (kgC/m2/d) */
-	public int deadstemc_storage_to_deadstemc_transfer; /* (kgC/m2/d) */
-	public int livecrootc_storage_to_livecrootc_transfer; /* (kgC/m2/d) */ /*
+	public double deadstemc_storage_to_deadstemc_transfer; /* (kgC/m2/d) */
+	public double livecrootc_storage_to_livecrootc_transfer; /* (kgC/m2/d) */ /*
 																			 * CHANGED
 																			 * TO
 																			 * INT
@@ -143,7 +143,7 @@ public class CFlux {
 																			 * PREVENT
 																			 * ERROR
 																			 */
-	public int deadcrootc_storage_to_deadcrootc_transfer; /* (kgC/m2/d) */ /*
+	public double deadcrootc_storage_to_deadcrootc_transfer; /* (kgC/m2/d) */ /*
 																			 * CHANGED
 																			 * TO
 																			 * INT
@@ -151,7 +151,7 @@ public class CFlux {
 																			 * PREVENT
 																			 * ERROR
 																			 */
-	public int gresp_storage_to_gresp_transfer; /* (kgC/m2/d) */ /*
+	public double gresp_storage_to_gresp_transfer; /* (kgC/m2/d) */ /*
 																	 * CHANGED TO
 																	 * INT TO
 																	 * PREVENT
@@ -171,49 +171,49 @@ public class CFlux {
 																 */
 
 	public CFlux(double dM_leafc_to_litr1c, double dM_leafc_to_litr2c, double dM_leafc_to_litr3c,
-			double dM_leafc_to_litr4c, double dM_frootc_to_litr1c, double dM_frootc_to_litr2c,
-			double dM_frootc_to_litr3c, double dM_frootc_to_litr4c, double dM_leafc_storage_to_litr1c,
-			double dM_frootc_storage_to_litr1c, double dM_livestemc_storage_to_litr1c,
-			double dM_deadstemc_storage_to_litr1c, double dM_livecrootc_storage_to_litr1c,
-			double dM_deadcrootc_storage_to_litr1c, double dM_leafc_transfer_to_litr1c,
-			double dM_frootc_transfer_to_litr1c, double dM_livestemc_transfer_to_litr1c,
-			double dM_deadstemc_transfer_to_litr1c, double dM_livecrootc_transfer_to_litr1c,
-			double dM_deadcrootc_transfer_to_litr1c, double dM_livestemc_to_cwdc, double dM_deadstemc_to_cwdc,
-			double dM_livecrootc_to_cwdc, double dM_deadcrootc_to_cwdc, double dM_gresp_storage_to_litr1c,
-			double dM_gresp_transfer_to_litr1c, double dM_leafc_to_fire, double dM_frootc_to_fire,
-			double dM_leafc_storage_to_fire, double dM_frootc_storage_to_fire, double dM_livestemc_storage_to_fire,
-			double dM_deadstemc_storage_to_fire, double dM_livecrootc_storage_to_fire,
-			double dM_deadcrootc_storage_to_fire, double dM_leafc_transfer_to_fire, double dM_frootc_transfer_to_fire,
-			double dM_livestemc_transfer_to_fire, double dM_deadstemc_transfer_to_fire,
-			double dM_livecrootc_transfer_to_fire, double dM_deadcrootc_transfer_to_fire, double dM_livestemc_to_fire,
-			double dM_deadstemc_to_fire, double dM_livecrootc_to_fire, double dM_deadcrootc_to_fire,
-			double dM_gresp_storage_to_fire, double dM_gresp_transfer_to_fire, double dM_litr1c_to_fire,
-			double dM_litr2c_to_fire, double dM_litr3c_to_fire, double dM_litr4c_to_fire, double dM_cwdc_to_fire,
-			double dLeafc_transfer_to_leafc, double dFrootc_transfer_to_frootc, double dLivestemc_transfer_to_livestemc,
-			double dDeadstemc_transfer_to_deadstemc, double dLivecrootc_transfer_to_livecrootc,
-			double dDeadcrootc_transfer_to_deadcrootc, double dLeafc_to_litr1c, double dLeafc_to_litr2c,
-			double dLeafc_to_litr3c, double dLeafc_to_litr4c, double dFrootc_to_litr1c, double dFrootc_to_litr2c,
-			double dFrootc_to_litr3c, double dFrootc_to_litr4c, double dLeaf_day_mr, double dLeaf_night_mr,
-			double dFroot_mr, double dLivestem_mr, double dLivecroot_mr, double dPsnsun_to_cpool,
-			double dPsnshade_to_cpool, double dCwdc_to_litr2c, double dCwdc_to_litr3c, double dCwdc_to_litr4c,
-			double dLitr1_hr, double dLitr1c_to_soil1c, double dLitr2_hr, double dLitr2c_to_soil2c,
-			double dLitr3c_to_litr2c, double dLitr4_hr, double dLitr4c_to_soil3c, double dSoil1_hr,
-			double dSoil1c_to_soil2c, double dSoil2_hr, double dSoil2c_to_soil3c, double dSoil3_hr,
-			double dSoil3c_to_soil4c, double dSoil4_hr, double dCpool_to_leafc, double dCpool_to_leafc_storage,
-			double dCpool_to_frootc, double dCpool_to_frootc_storage, double dCpool_to_livestemc,
-			double dCpool_to_livestemc_storage, double dCpool_to_deadstemc, double dCpool_to_deadstemc_storage,
-			double dCpool_to_livecrootc, double dCpool_to_livecrootc_storage, double dCpool_to_deadcrootc,
-			double dCpool_to_deadcrootc_storage, double dCpool_to_gresp_storage, double dCpool_leaf_gr,
-			double dCpool_leaf_storage_gr, double dTransfer_leaf_gr, double dCpool_froot_gr,
-			double dCpool_froot_storage_gr, double dTransfer_froot_gr, double dCpool_livestem_gr,
-			double dCpool_livestem_storage_gr, double dTransfer_livestem_gr, double dCpool_deadstem_gr,
-			double dCpool_deadstem_storage_gr, double dTransfer_deadstem_gr, double dCpool_livecroot_gr,
-			double dCpool_livecroot_storage_gr, double dTransfer_livecroot_gr, double dCpool_deadcroot_gr,
-			double dCpool_deadcroot_storage_gr, double dTransfer_deadcroot_gr, double dLeafc_storage_to_leafc_transfer,
-			double dFrootc_storage_to_frootc_transfer, double dLivestemc_storage_to_livestemc_transfer,
-			int dDeadstemc_storage_to_deadstemc_transfer, int dLivecrootc_storage_to_livecrootc_transfer,
-			int dDeadcrootc_storage_to_deadcrootc_transfer, int dGresp_storage_to_gresp_transfer,
-			double dLivestemc_to_deadstemc, double dLivecrootc_to_deadcrootc) {
+				 double dM_leafc_to_litr4c, double dM_frootc_to_litr1c, double dM_frootc_to_litr2c,
+				 double dM_frootc_to_litr3c, double dM_frootc_to_litr4c, double dM_leafc_storage_to_litr1c,
+				 double dM_frootc_storage_to_litr1c, double dM_livestemc_storage_to_litr1c,
+				 double dM_deadstemc_storage_to_litr1c, double dM_livecrootc_storage_to_litr1c,
+				 double dM_deadcrootc_storage_to_litr1c, double dM_leafc_transfer_to_litr1c,
+				 double dM_frootc_transfer_to_litr1c, double dM_livestemc_transfer_to_litr1c,
+				 double dM_deadstemc_transfer_to_litr1c, double dM_livecrootc_transfer_to_litr1c,
+				 double dM_deadcrootc_transfer_to_litr1c, double dM_livestemc_to_cwdc, double dM_deadstemc_to_cwdc,
+				 double dM_livecrootc_to_cwdc, double dM_deadcrootc_to_cwdc, double dM_gresp_storage_to_litr1c,
+				 double dM_gresp_transfer_to_litr1c, double dM_leafc_to_fire, double dM_frootc_to_fire,
+				 double dM_leafc_storage_to_fire, double dM_frootc_storage_to_fire, double dM_livestemc_storage_to_fire,
+				 double dM_deadstemc_storage_to_fire, double dM_livecrootc_storage_to_fire,
+				 double dM_deadcrootc_storage_to_fire, double dM_leafc_transfer_to_fire, double dM_frootc_transfer_to_fire,
+				 double dM_livestemc_transfer_to_fire, double dM_deadstemc_transfer_to_fire,
+				 double dM_livecrootc_transfer_to_fire, double dM_deadcrootc_transfer_to_fire, double dM_livestemc_to_fire,
+				 double dM_deadstemc_to_fire, double dM_livecrootc_to_fire, double dM_deadcrootc_to_fire,
+				 double dM_gresp_storage_to_fire, double dM_gresp_transfer_to_fire, double dM_litr1c_to_fire,
+				 double dM_litr2c_to_fire, double dM_litr3c_to_fire, double dM_litr4c_to_fire, double dM_cwdc_to_fire,
+				 double dLeafc_transfer_to_leafc, double dFrootc_transfer_to_frootc, double dLivestemc_transfer_to_livestemc,
+				 double dDeadstemc_transfer_to_deadstemc, double dLivecrootc_transfer_to_livecrootc,
+				 double dDeadcrootc_transfer_to_deadcrootc, double dLeafc_to_litr1c, double dLeafc_to_litr2c,
+				 double dLeafc_to_litr3c, double dLeafc_to_litr4c, double dFrootc_to_litr1c, double dFrootc_to_litr2c,
+				 double dFrootc_to_litr3c, double dFrootc_to_litr4c, double dLeaf_day_mr, double dLeaf_night_mr,
+				 double dFroot_mr, double dLivestem_mr, double dLivecroot_mr, double dPsnsun_to_cpool,
+				 double dPsnshade_to_cpool, double dCwdc_to_litr2c, double dCwdc_to_litr3c, double dCwdc_to_litr4c,
+				 double dLitr1_hr, double dLitr1c_to_soil1c, double dLitr2_hr, double dLitr2c_to_soil2c,
+				 double dLitr3c_to_litr2c, double dLitr4_hr, double dLitr4c_to_soil3c, double dSoil1_hr,
+				 double dSoil1c_to_soil2c, double dSoil2_hr, double dSoil2c_to_soil3c, double dSoil3_hr,
+				 double dSoil3c_to_soil4c, double dSoil4_hr, double dCpool_to_leafc, double dCpool_to_leafc_storage,
+				 double dCpool_to_frootc, double dCpool_to_frootc_storage, double dCpool_to_livestemc,
+				 double dCpool_to_livestemc_storage, double dCpool_to_deadstemc, double dCpool_to_deadstemc_storage,
+				 double dCpool_to_livecrootc, double dCpool_to_livecrootc_storage, double dCpool_to_deadcrootc,
+				 double dCpool_to_deadcrootc_storage, double dCpool_to_gresp_storage, double dCpool_leaf_gr,
+				 double dCpool_leaf_storage_gr, double dTransfer_leaf_gr, double dCpool_froot_gr,
+				 double dCpool_froot_storage_gr, double dTransfer_froot_gr, double dCpool_livestem_gr,
+				 double dCpool_livestem_storage_gr, double dTransfer_livestem_gr, double dCpool_deadstem_gr,
+				 double dCpool_deadstem_storage_gr, double dTransfer_deadstem_gr, double dCpool_livecroot_gr,
+				 double dCpool_livecroot_storage_gr, double dTransfer_livecroot_gr, double dCpool_deadcroot_gr,
+				 double dCpool_deadcroot_storage_gr, double dTransfer_deadcroot_gr, double dLeafc_storage_to_leafc_transfer,
+				 double dFrootc_storage_to_frootc_transfer, double dLivestemc_storage_to_livestemc_transfer,
+				 double dDeadstemc_storage_to_deadstemc_transfer, double dLivecrootc_storage_to_livecrootc_transfer,
+				 double dDeadcrootc_storage_to_deadcrootc_transfer, double dGresp_storage_to_gresp_transfer,
+				 double dLivestemc_to_deadstemc, double dLivecrootc_to_deadcrootc) {
 
 		this.m_leafc_to_litr1c = dM_leafc_to_litr1c;
 		this.m_leafc_to_litr2c = dM_leafc_to_litr2c;
